+++ HISTORY.rst
@@ -2,7 +2,8 @@
 ------------------
 
 - Add 'copy' functionality to update_object endpoint. In this case the view accepts the (augeias) location
-of the file object and copies it to a new location. (#17)
+of the file object (host url, collection_key, container_key and object_key) and copies it to a new location
+on the same augeias instance. (#17)
 - Add Endpoint to create object and key. (#20)
 - Add Endpoint to retrieve meta data (mimetype, size, time last modification) from object.